@@ -4,14 +4,14 @@ import { CartItem } from "../restaurant-detail/shopping-cart/item-cart.model";
 import { ShoppingCartService } from "../restaurant-detail/shopping-cart/shopping-cart.service";
 import { Order } from "./order.model";
 import 'rxjs/add/operator/map'
-import { Http, Headers, RequestOptions } from "@angular/http";
+import { HttpClient} from "@angular/common/http";
 import { MEAT_API } from "../app.api";
 
 @Injectable()
 
 export class OrderService{
 
-    constructor(private cartService: ShoppingCartService, private http: Http){}
+    constructor(private cartService: ShoppingCartService, private http: HttpClient){}
 
     itemsValue():number{
         return this.cartService.total()
@@ -35,12 +35,7 @@ export class OrderService{
     }
 
     checkOrder(order: Order): Observable<string>{
-        const headers = new Headers();
-        headers.append('Content-Type', 'Application/json')
-        return this.http.post(`${MEAT_API}/orders`, 
-                                JSON.stringify(order),
-                                new RequestOptions({headers: headers}))
-                        .map(response => response.json())
+        return this.http.post<Order>(`${MEAT_API}/orders`,order)
                         .map(order => order.id)
     }
 }
